Add `Partners` module
This adds a second kind of assignment task. Each assignment task will
now be split out into its own sub-module, and users can pick which
namespace they want for their application.
