+++ Project.toml
@@ -6,6 +6,7 @@ version = "1.0.0"
 [deps]
 HiGHS = "87dc4568-4c63-4d18-b0c0-bb2238e4078b"
 JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
+Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 
 [weakdeps]
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
@@ -17,10 +18,12 @@ AssignGroupsCSV = "CSV"
 CSV = "0.10"
 HiGHS = "1"
 JuMP = "1"
+Statistics = "1"
 julia = "1.9"
 
 [extras]
+CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Test"]
+test = ["CSV", "Test"]
