+++ test/csvfiles/partners.csv
@@ -1,0 +1,7 @@
+First,Last,Score,Partners
+StudentA,Last,1.0,"StudentD Last"
+StudentB,Last,2.0,
+StudentC,Last,3.0,
+StudentD,Last,1.0,
+StudentE,Last,2.0,
+StudentF,Last,3.0,
